Correctly treat "null-like" IDs
It turns out that if an attribute is set via:

  elem.setAttribute(attrName, null);

then, when retrieving that attribute, the value doesn't come back as null,
but as the _string_ "null". This is what happens in the case of React and
JSX apparently, so we'll assume that no one is setting the value to be the
string "null" deliberately.